Consolidate test and helper files
