+++ package.json
@@ -14,8 +14,8 @@
     "build": "rollup -c",
     "prepublishOnly": "npm test",
     "test": "npm run test:src && npm run lint",
-    "test:src": "mocha --require esm --recursive test",
-    "test:bundle": "npm run build && mocha --recursive test-bundle",
+    "test:src": "mocha --require esm 'test/**/*.test.js'",
+    "test:bundle": "npm run build && mocha 'test/**/*.test-bundle.js'",
     "lint": "standard --env=mocha --env=worker --verbose",
     "lint:fix": "standard --env=mocha --env=worker --fix --verbose"
   },
